Update requirements and remove Flask-Script
* Move script commands to app.py (for use with Flask-CLI)
* Remove manager.py
* Update readme to reflect dockerization
* Update docker file
* Update requirements
