+++ app/config.py
@@ -4,7 +4,7 @@ DEBUG = os.environ.get('DEBUG', True)
 REDIS_URL = os.environ.get('REDIS_URL', 'redis://redis:6379/')
 RQ_REDIS_URL = REDIS_URL
 
-DATABASE_URL = os.environ.get('DATABASE_URL', 'postgres://localhost:5432/votebot-forms')
+DATABASE_URL = os.environ.get('DATABASE_URL', 'postgres://postgres:password@db/votebot-forms')
 VOTEBOT_API_KEY = os.environ.get('VOTEBOT_API_KEY', None)
 
 AWS_ACCESS_KEY = os.environ.get('AWS_ACCESS_KEY')
