+++ docker-compose.yml
@@ -5,12 +5,16 @@ services:
     build: .
     ports:
       - "5000:5000"
-    command: python manager.py runserver --host 0.0.0.0
+    command: flask run --host 0.0.0.0
     environment:
       - EASYPOST_API_KEY=[YOUR_EASYPOST_API_KEY]
       - LOB_API_KEY=[YOUR_LOB_API_KEY]
       - AWS_ACCESS_KEY=[YOUR_AWS_ACCESS_KEY]
       - AWS_SECRET_KEY=[YOUR_AWS_SECRET_KEY]
+      - SENTRY_DSN=[YOUR_SENTRY_DSN]
+      - SMARTY_STREETS_AUTH_ID=[YOUR_SMARTY_STREETS_AUTH_ID]
+      - SMARTY_STREETS_AUTH_TOKEN=[YOUR_SMARTY_STREETS_AUTH_TOKEN]
+      - DEBUG_SUBMIT=TRUE
     volumes:
       - .:/usr/src/app
     depends_on:
@@ -19,19 +23,23 @@ services:
 
   worker:
     build: .
-    command: python manager.py rq worker 
+    command: flask rq worker
     environment:
+      - FLASK_APP=app/wsgi.py
       - AWS_ACCESS_KEY=[YOUR_AWS_ACCESS_KEY]
       - AWS_SECRET_KEY=[YOUR_AWS_SECRET_KEY]
     volumes:
       - .:/usr/src/app
     depends_on:
       - redis
+      - db
 
   redis:
     image: redis:3-alpine
 
   db:
-    image: postgres
-    ports:
-    - "5432:5432"
+    image: postgres:10.7-alpine
+    environment:
+      - POSTGRES_PASSWORD=[POSTGRES_PASSWORD]
+      - POSTGRES_USER=postgres
+      - POSTGRES_DB=votebot-forms
