+++ requirements/common.txt
@@ -1,9 +1,8 @@
-Flask==0.11.1
-Flask-RQ2==17.0
-Flask-Script==2.0.5
-Jinja2==2.8
+Flask==1.0.2
+Flask-RQ2[cli]==18.3
+Jinja2==2.10
 MarkupSafe==0.23
-Werkzeug==0.11.10
+Werkzeug==0.14
 beautifulsoup4==4.4.1
 blinker==1.4
 boto==2.42.0
@@ -22,10 +21,10 @@ psycopg2==2.7.3.2
 pyOpenSSL==17.5.0
 pyasn1==0.3.6
 python-dateutil==2.5.3
-redis>=2.7.0
-requests==2.18.4
+redis==3.2.1
+requests==2.21
 rq-scheduler>=0.6.1
-rq>=0.6.0,<0.7.0
+rq==0.13.0
 robobrowser==0.5.3
 six==1.10.0
 smartystreets.py==0.4.0
